Agent2/requirements.txt
google-generativeai>=0.7.0
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
